Disallow placeholder recipe discovering
